teams-install-proto-llm-full (1)/teams-install-proto-llm-full/requirements.txt
streamlit==1.24.0
sqlalchemy==2.0.19
pandas==2.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
click==8.1.7
huggingface_hub==0.18.1
pymysql
